+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## WIP
+- Fix sidebar heading colour contrast
+
+
 ## 0.0.16
 - Add option for inlined toolbar save buttons
 - Disable sidebar menu on edit views